Added lots more javadoc, fixed solarLint warnings, adjusted the action hierarchy by removing duplicate code and adjusted the canExecute logic.
