+++ eisbw/src/main/java/eisbw/configuration/BooleanString.java
@@ -1,8 +1,18 @@
 package eisbw.configuration;
 
+/**
+ * @author Danny & Harm - A boolean as a string.
+ *
+ */
 public class BooleanString {
   String data;
-  
+
+  /**
+   * The BooleanString constructor.
+   * 
+   * @param data
+   *          The boolean string.
+   */
   public BooleanString(String data) {
     this.data = data;
   }
