+++ eisbw/src/main/java/eisbw/configuration/RaceString.java
@@ -1,8 +1,18 @@
 package eisbw.configuration;
 
+/**
+ * @author Danny & Harm - A race as a string.
+ *
+ */
 public class RaceString {
   String data;
-  
+
+  /**
+   * The RaceString constructor.
+   * 
+   * @param data
+   *          The race string.
+   */
   public RaceString(String data) {
     this.data = data;
   }
