+++ eisbw/src/main/java/eisbw/constants/Races.java
@@ -3,32 +3,36 @@ package eisbw.constants;
 import java.util.ArrayList;
 import java.util.List;
 
+/**
+ * @author Danny & Harm - The data class with all race definition input.
+ *
+ */
 public class Races {
 
-  private static List<String> races;
+  private static List<String> raceNames;
 
   private Races() {
-    //Static class.
+    // Static class.
   }
-  
+
   /**
    * Get a list from all racenames
    * 
    * @return racenames.
    */
   public static List<String> getRaceList() {
-    if (races == null) {
-      races = new ArrayList<>(7);
-      races.add("terran");
-      races.add("protoss");
-      races.add("zerg");
-      races.add("random");
-      races.add("randomtp");
-      races.add("randomtz");
-      races.add("randompz");
-      races.add("test");
+    if (raceNames == null) {
+      raceNames = new ArrayList<>(7);
+      raceNames.add("terran");
+      raceNames.add("protoss");
+      raceNames.add("zerg");
+      raceNames.add("random");
+      raceNames.add("randomtp");
+      raceNames.add("randomtz");
+      raceNames.add("randompz");
+      raceNames.add("test");
     }
-    return races;
+    return raceNames;
   }
 
 }
