+++ eisbw/src/main/java/eisbw/percepts/MapPercept.java
@@ -11,6 +11,14 @@ public class MapPercept extends Percept {
 
   private static final long serialVersionUID = 1L;
 
+  /**
+   * The MapPercept constructor.
+   * 
+   * @param width
+   *          The width of the map
+   * @param height
+   *          The height of the map
+   */
   public MapPercept(int width, int height) {
     super(Percepts.MAP, new Numeral(width), new Numeral(height));
   }
