+++ eisbw/src/main/java/eisbw/percepts/QueueSizePercept.java
@@ -11,6 +11,12 @@ public class QueueSizePercept extends Percept {
 
   private static final long serialVersionUID = 1L;
 
+  /**
+   * The QueueSize constructor.
+   * 
+   * @param queueSize
+   *          The size of the queue.
+   */
   public QueueSizePercept(int queueSize) {
     super(Percepts.QUEUESIZE, new Numeral(queueSize));
   }
