+++ eisbw/src/main/java/eisbw/percepts/RallyUnitPercept.java
@@ -11,6 +11,12 @@ public class RallyUnitPercept extends Percept {
 
   private static final long serialVersionUID = 1L;
 
+  /**
+   * The RallyUnitPercept constructor.
+   * 
+   * @param id
+   *          The id of the unit.
+   */
   public RallyUnitPercept(int id) {
     super(Percepts.RALLYUNIT, new Numeral(id));
   }
