+++ eisbw/src/main/java/eisbw/percepts/ResourcesPercept.java
@@ -11,6 +11,18 @@ public class ResourcesPercept extends Percept {
 
   private static final long serialVersionUID = 1L;
 
+  /**
+   * The ResourcesPercept constructor.
+   * 
+   * @param minerals
+   *          The amount of collected minerals.
+   * @param gas
+   *          The amount of collected gas.
+   * @param csupply
+   *          The current amount of supply.
+   * @param tsupply
+   *          The total amount of supply.
+   */
   public ResourcesPercept(int minerals, int gas, int csupply, int tsupply) {
     super(Percepts.RESOURCES, new Numeral(minerals), new Numeral(gas), new Numeral(csupply),
         new Numeral(tsupply));
