+++ eisbw/src/main/java/eisbw/percepts/SpaceProvidedPercept.java
@@ -11,6 +11,14 @@ public class SpaceProvidedPercept extends Percept {
 
   private static final long serialVersionUID = 1L;
 
+  /**
+   * The SpaceProvidedPercept.
+   * 
+   * @param used
+   *          The already used amount of space.
+   * @param max
+   *          The maximum amount of space available.
+   */
   public SpaceProvidedPercept(int used, int max) {
     super(Percepts.SPACEPROVIDED, new Numeral(used), new Numeral(max));
   }
