+++ eisbw/src/main/java/eisbw/percepts/UpgradePercept.java
@@ -11,6 +11,12 @@ public class UpgradePercept extends Percept {
 
   private static final long serialVersionUID = 1L;
 
+  /**
+   * The UpgradePercept constructor.
+   * 
+   * @param upgrade
+   *          The name of the upgrade type.
+   */
   public UpgradePercept(String upgrade) {
     super(Percepts.UPGRADING, new Identifier(upgrade));
   }
