+++ eisbw/src/main/java/eisbw/percepts/WorkerActivityPercept.java
@@ -12,6 +12,14 @@ public class WorkerActivityPercept extends Percept {
 
   private static final long serialVersionUID = 1L;
 
+  /**
+   * The WorkerActivityPercept constructor.
+   * 
+   * @param id
+   *          The id of te worker unit.
+   * @param activity
+   *          The current activity of the worker unit.
+   */
   public WorkerActivityPercept(int id, String activity) {
     super(Percepts.WORKERACTIVITY, new Numeral(id), new Identifier(activity));
   }
